section 1.4: create homeview model and homeview refractor
