+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Abhinav Budhwar on 27/1/25.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+}
