+++ FitnessApp/Home/ViewModels/HomeViewModel.swift
@@ -1,0 +1,31 @@
+//
+//  HomeViewModel.swift
+//  FitnessApp
+//
+//  Created by Abhinav Budhwar on 27/1/25.
+//
+
+import Foundation
+class HomeViewModel: ObservableObject {
+    
+    var calories: Int = 123
+    var active: Int = 52
+    var stand: Int = 8
+    
+    
+    var mockActivites = [
+        Activity(id: 0, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .green, amount: "9812"),
+        Activity(id: 1, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .red, amount: "9812"),
+        Activity(id: 2, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .blue, amount: "9812"),
+        Activity(id: 3, title: "Today steps", subtitle: "Goal 12,000", image: "figure.run", tintColor: .purple, amount: "104,812"),
+    ]
+    
+    var mockWorkouts = [
+        Workout(id: 0, title: "Running", image: "figure.run", tintColor: .cyan, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
+        Workout(id: 1, title: "Running", image: "figure.run", tintColor: .red, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
+        Workout(id: 2, title: "Running", image: "figure.run", tintColor: .purple, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
+        Workout(id: 3, title: "Running", image: "figure.run", tintColor: .green, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
+    ]
+     
+    
+}
